# Improve performance of re-assigning queriers to tenants in scheduler (aka shuffle sharding)