**What this PR does / why we need it**:

Whenever a querier connects or disconnects to the scheduler, the queue blocks and re-assigns queriers to tenants. This is known as shuffle sharding.

However, when auto-scaling of queriers is enabled, it happens, that a lot of queriers are added at the same time, causing the queue to block and sub-queries to wait.

Unfortunately, the re-distribution of queriers was relatively slow. One part of the slowness is caused by generating random numbers using `math/rand`.

This PR replaces the stdlib `math/rand` with `pgregory.net/rand` and uses `SwissMap` instead of the stdlib `map` for the tenant-querier-mapping.


**Special notes for your reviewer**:

```
$ benchstat scheduler-old.txt scheduler-rand.txt scheduler-swissmap.txt 
name \ time/op                              scheduler-old.txt  scheduler-rand.txt  scheduler-swissmap.txt
RecomputeTenantQueriers/from_100_to_500-8          474ms ± 0%          200ms ± 0%              143ms ± 0%
RecomputeTenantQueriers/from_200_to_500-8          457ms ± 0%          228ms ± 0%              148ms ± 0%
RecomputeTenantQueriers/from_400_to_500-8          195ms ± 0%           92ms ± 0%               64ms ± 0%
RecomputeTenantQueriers/from_500_to_1000-8         1.60s ± 0%          0.91s ± 0%              0.65s ± 0%

name \ queriers                             scheduler-old.txt  scheduler-rand.txt  scheduler-swissmap.txt
RecomputeTenantQueriers/from_100_to_500-8            400 ± 0%            400 ± 0%                400 ± 0%
RecomputeTenantQueriers/from_200_to_500-8            300 ± 0%            300 ± 0%                300 ± 0%
RecomputeTenantQueriers/from_400_to_500-8            100 ± 0%            100 ± 0%                100 ± 0%
RecomputeTenantQueriers/from_500_to_1000-8           500 ± 0%            500 ± 0%                500 ± 0%
```

**Checklist**
- [ ] Reviewed the [`CONTRIBUTING.md`](https://github.com/grafana/loki/blob/main/CONTRIBUTING.md) guide (**required**)
- [ ] Documentation added
- [ ] Tests updated
- [ ] `CHANGELOG.md` updated
- [ ] Changes that require user attention or interaction to upgrade are documented in `docs/sources/upgrading/_index.md`